move build to github; refactor; java 17; update gradle

+++ build_docs.sh
@@ -1,6 +1,14 @@
 #!/bin/sh
 
+here="$(dirname "$(readlink -f "$0")")"
+cd "$here" || exit 1
+
 h=docs/examples.html
+if [ -e $h ] ; then
+    echo "ERROR: file $h exists; will not overwrite."
+    exit 1
+fi
+
 echo "<!doctype html><html><body>" >$h
 for a in docs/*.afl ; do
     ./gradlew run --args="--output=$a.dot $a"
